Добавил рандомную генерацию последовательности вопросов, вернул функцию flash_cards в bot.py из-за того, что хэндлер сообщений не ловит их в другом файле при применении функции bot.register_next_step_handler()
